refactor: use new solid icon wrapper component import path to solid icon components
